change place of comments functional
