+++ murr_card/models.py
@@ -33,17 +33,6 @@ class MurrCard(models.Model):
                 img.save(self.cover.path, 'webp')
 
 
-class Comment(models.Model):
-    """ Модель комментнария для murr_card """
-    murr_card = models.ForeignKey(MurrCard, on_delete=models.CASCADE, related_name='comments', verbose_name='murr_card')
-    owner = models.ForeignKey(Murren, on_delete=models.CASCADE, related_name='comments',
-                              verbose_name='Автор комментария')
-    text = models.TextField(max_length=512)
-
-    def __str__(self):
-        return f'{self.owner}: "{self.text}"'
-
-
 class EditorImageForMurrCard(models.Model):
     murr_editor_image = models.ImageField(upload_to='editor_image_for_murr_card/%Y/%m/%d/', null=True, max_length=255)
