+++ murr_card/serializers.py
@@ -1,14 +1,7 @@
 from rest_framework import serializers
 
-from .models import MurrCard, EditorImageForMurrCard, Comment
-
-
-class CommentSerializer(serializers.ModelSerializer):
-    owner_name = serializers.ReadOnlyField(source='owner.username')
-
-    class Meta:
-        model = Comment
-        fields = ('owner', 'murr_card', 'text', 'owner_name')
+from comments.serializers import CommentSerializer
+from .models import MurrCard, EditorImageForMurrCard
 
 
 class MurrCardSerializers(serializers.ModelSerializer):
